+++ editor/tsconfig.json
@@ -3,10 +3,15 @@
         "target": "es5",
         "noImplicitAny": true,
         "noImplicitReturns": true,
+        "noImplicitThis": true,
         "declaration": true,
-        "out": "../built/editor.js",
+        "module": "commonjs",
+        "moduleResolution": "node",
+        "isolatedModules": false,
+        "outDir": "../built/editor",
         "rootDir": ".",
         "newLine": "LF",
-        "sourceMap": false
+        "sourceMap": false,
+        "jsx": "react"
     }
 }
